[step1] refactor: issue components 분리
